Shift common announcement methods into a concern
