+++ app/models/generic_announcement.rb
@@ -3,10 +3,9 @@
 #
 # This is the abstract single-table inheritance table used for both announcement types.
 class GenericAnnouncement < ApplicationRecord
-  after_initialize :set_defaults, if: :new_record?
+  include AnnouncementConcern
+
   acts_as_readable on: :updated_at
-  after_create :mark_as_read_by_creator
-  after_update :mark_as_read_by_updater
 
   belongs_to :instance, inverse_of: :announcements
 
@@ -29,23 +28,5 @@ class GenericAnnouncement < ApplicationRecord
 
   def sticky?
     false
-  end
-
-  private
-
-  # Set default values
-  def set_defaults
-    self.start_at ||= Time.zone.now
-    self.end_at ||= 7.days.from_now
-  end
-
-  # Mark announcement as read for the creator
-  def mark_as_read_by_creator
-    mark_as_read! for: creator
-  end
-
-  # Mark announcement as read for the updater
-  def mark_as_read_by_updater
-    mark_as_read! for: updater
   end
 end
